[Refactor] Move database interactions to driver class
This makes it easier to modify the repository's interaction with
the database. Rather than having multiple repository classes,
the repository is now provided with a driver that it uses for its
base interactions with the database (via Eloquent builders). This
means that the repository's behaviour can be overridden by
writing new driver classes. The soft deletes feature is one such
instance where the feature can be implemented by writing the
soft-deletes database interactions in a driver class, rather than
a separate repository.