[Refactor] Update query builder implementations
This allows logic from within the `JsonApiBuilder` class to be
executed in different locations. For example, the filter logic
can now be used from within the `WhereHas` and `WhereDoesntHave`
filter classes.

As part of this refactor, a new `QueryBuilder` namespace was added,
with all classes that are to do with applying query builder logic
moved into that namespace.